JobPortal/requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
psycopg2-binary==2.9.10
PyJWT==2.12.1
sqlparse==0.5.5
tzdata==2025.3
